requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
pyjwt==2.1.0
werkzeug==2.0.1
torch==1.9.0
email-validator==1.1.3